Feat/update terraform tab (#1)
* chore: Update costEstimate function in actions.ts to include specific detail in the configuration summary

* chore: Update CostTab component to display a skeleton table with mock columns

* chore: Update ExplainSection and ChatControl components

Refactor ExplainSection component to add the "overflow-y-hidden" class to the CustomTabsContent component for the "terraform" value. This improves the UI by hiding the overflow of content.
Update ChatControl component to remove the commented out code and add a new function checkTerraformSync for synchronous file upload and logging. This enhances the functionality of the component.